Feature/core upd (#20)
* Added hide immediately for uimanager

* Fix App.Restart() issue

In TRS I found miss refs (UIManager) in processable that break all game. When we restart game we will load new scene, but if we don't wait to fully loaded scene and invoke Init(manifest) it can grab UIManager, get all windows and add all this staff to App processables.